feat: update driver entity and props
